requirements/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
fastapi==0.95.1
frozenlist==1.3.3
h11==0.14.0
idna==3.4
multidict==6.0.4
openai==0.27.4
pydantic==1.10.7
python-dotenv==1.0.0
requests==2.28.2
sniffio==1.3.0
starlette==0.26.1
tqdm==4.65.0
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.21.1
yarl==1.8.2
motor==2.5.1
